resilient-decision-system/rds-export/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
jsonschema==4.23.0
pyyaml==6.0.2
redis==5.1.1
httpx==0.27.2
pytest==8.3.3
pytest-httpx==0.30.0
